+++ apps/explorer/test/explorer/chain/import/runner/internal_transactions_test.exs
@@ -22,6 +22,74 @@ defmodule Explorer.Chain.Import.Runner.InternalTransactionsTest do
       assert :error == Repo.get(Transaction, transaction.hash).status
     end
 
+    test "transaction's has_error_in_iternal_txs become true when its internal_transaction (where index != 0) has an error" do
+      transaction = insert(:transaction) |> with_block(status: :ok)
+      insert(:pending_block_operation, block_hash: transaction.block_hash, fetch_internal_transactions: true)
+
+      assert :ok == transaction.status
+      assert nil == transaction.has_error_in_iternal_txs
+
+      index = 0
+      error = nil
+
+      internal_transaction_changes = make_internal_transaction_changes(transaction, index, error)
+
+      index = 1
+      error = "Reverted"
+
+      internal_transaction_changes_1 = make_internal_transaction_changes(transaction, index, error)
+
+      assert {:ok, _} = run_internal_transactions([internal_transaction_changes, internal_transaction_changes_1])
+      tx = Repo.get(Transaction, transaction.hash)
+
+      assert :ok == tx.status
+      assert true == tx.has_error_in_iternal_txs
+    end
+
+    test "transaction's has_error_in_iternal_txs become false when its internal_transaction (where index == 0) has an error" do
+      transaction = insert(:transaction) |> with_block(status: :ok)
+      insert(:pending_block_operation, block_hash: transaction.block_hash, fetch_internal_transactions: true)
+
+      assert :ok == transaction.status
+      assert nil == transaction.has_error_in_iternal_txs
+
+      index = 0
+      error = "Reverted"
+
+      internal_transaction_changes = make_internal_transaction_changes(transaction, index, error)
+
+      assert {:ok, _} = run_internal_transactions([internal_transaction_changes])
+      tx = Repo.get(Transaction, transaction.hash)
+
+      assert :error == tx.status
+      assert false == tx.has_error_in_iternal_txs
+    end
+
+    test "transaction's has_error_in_iternal_txs become false when its internal_transaction has no error" do
+      transaction = insert(:transaction) |> with_block(status: :ok)
+      insert(:pending_block_operation, block_hash: transaction.block_hash, fetch_internal_transactions: true)
+
+      assert :ok == transaction.status
+      assert nil == transaction.has_error_in_iternal_txs
+
+      index = 0
+      error = nil
+
+      internal_transaction_changes = make_internal_transaction_changes(transaction, index, error)
+
+      index = 1
+      error = nil
+
+      internal_transaction_changes_1 = make_internal_transaction_changes(transaction, index, error)
+
+      assert {:ok, _} = run_internal_transactions([internal_transaction_changes, internal_transaction_changes_1])
+
+      tx = Repo.get(Transaction, transaction.hash)
+
+      assert :ok == tx.status
+      assert false == tx.has_error_in_iternal_txs
+    end
+
     test "simple coin transfer's status becomes :error when its internal_transaction has an error" do
       transaction = insert(:transaction) |> with_block(status: :ok)
       insert(:pending_block_operation, block_hash: transaction.block_hash, fetch_internal_transactions: true)